% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyph.support-fr.R
\name{hyph.support.fr}
\alias{hyph.support.fr}
\title{Language support for French}
\usage{
hyph.support.fr()
}
\description{
This function adds support for French to the sylly package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular,
      a new set of hyphenation patterns is being added (see \code{\link{hyph.fr}}).
To use the patterns with \code{\link[sylly:hyphen]{hyphen}}, simply use the abbreviation:
\itemize{
  \item {\code{"fr"}} {--- French hyphenation patterns}
}
}
\seealso{
\code{\link[sylly:hyphen]{hyphen}},
   \code{\link[sylly:set.hyph.support]{set.hyph.support}}
}
